% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gss12_values}
\alias{gss12_values}
\title{General Social Survey 2012: Schwartz Values Module}
\format{
## `gss12_values`
A tibble with 1,255 rows and 21 columns:
\describe{
  \item{country}{Country name}
  \item{iso2, iso3}{2 & 3 letter ISO country codes}
  \item{year}{Year}
  ...
}
}
\source{
<https://www.icpsr.umich.edu/web/NADAC/studies/35478>
}
\usage{
gss12_values
}
\description{
The General Social Survey (GSS) collects information from the general public on a wide variety of subjects, including attitudes toward social issues, religion, education, jobs and the economy, government and other institutions, politics, and policy issues. The Schwartz Values module is a part of GSS 2012 questionnaire. This module is a good example of ordinal data, particularly Likert items.
}
\details{
The items are 6-level Likert scale from “Not like me at all” to “Very much like me”. Each question describes a person who adheres to certain values, the respondent is asked to what extent that person is like him.

*Module instruction:* “Now I will briefly describe some people. Please listen to each description and tell me how much each person is or is not like you. Use this card for your answer.”

*Possible responses:*
1.	Not like me at all
2.	Not like me
3.	A little like me
4.	Somewhat like me
5.	Like me
6.	Very much like me
}
\references{
Smith, Tom W., Marsden, Peter V., and Hout, Michael. General Social Survey, 2012 Merged Data, Including a Cultural Module [United States]. Inter-university Consortium for Political and Social Research [distributor], 2016-05-26. https://doi.org/10.3886/ICPSR35478.v4
}
\keyword{datasets}
