\name{polychoric-package}
\alias{polychoric-package}
\alias{polychoric}
\docType{package}
\title{\packageTitle{polychoric}}
\description{\packageDescription{polychoric}}
\section{Package Content}{\packageIndices{polychoric}}
\author{\packageAuthor{polychoric}}
\section{Maintainer}{\packageMaintainer{polychoric}}
\keyword{package}
